requirements.txt
Django
south
django-tastypie
pexpect
